+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -230,7 +230,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K96QTJRH1CG3GKY6RWS1YD0X"
+        "01K975TAY23BXKSC23TB0WM8R3"
     )
     r"""The unique identifier of the data source"""