+++ packages/orq-rc/src/orq_ai_sdk/models/getbudgetop.py
@@ -161,5 +161,5 @@ class GetBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-11-04T06:10:36.164Z")
+    updated: Optional[datetime] = parse_datetime("2025-11-04T10:15:08.760Z")
     r"""The date and time the resource was last updated"""