+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -132,7 +132,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K96QTJRE1TNQN1N5H2W4068M"
+        "01K975TAXZRQQDFBBS67MA5JVA"
     )
     r"""The unique identifier of the data source"""