+++ packages/orq-rc/RELEASES.md
@@ -3809,3 +3809,13 @@ Based on:
 - [python v4.0.0-rc.4] packages/orq-rc
 ### Releases
 - [PyPI v4.0.0-rc.4] https://pypi.org/project/orq-ai-sdk/4.0.0-rc.4 - packages/orq-rc
+
+## 2025-11-04 10:17:02
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.648.0 (2.737.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.0.0-rc.5] packages/orq-rc
+### Releases
+- [PyPI v4.0.0-rc.5] https://pypi.org/project/orq-ai-sdk/4.0.0-rc.5 - packages/orq-rc